backend/mythesisbackend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.6
django-cors-headers==4.3.0
django-osm-field==0.5.0
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
Pillow==10.1.0
ExifRead==3.0.0
django-q==1.3.9
requests
sendgrid>=6.0.0
